Refactor X-FILES/DANA into separate packages
