+++ src/main/scala/Control.scala
@@ -22,7 +22,7 @@ class ControlCacheInterfaceRespLearn(implicit p: Parameters)
   val globalWtptr = UInt(INPUT, 16) //[TODO] possibly fragile
 }
 
-class ControlCacheInterfaceReq(implicit p: Parameters) extends XFilesBundle()(p) {
+class ControlCacheInterfaceReq(implicit p: Parameters) extends DanaBundle()(p) {
   val request = UInt(width = log2Up(3)) // [TODO] fragile on Constants.scala
   val asid = UInt(width = asidWidth)
   val nnid = UInt(width = nnidWidth)
